backend/requirements.txt
# 核心框架
fastapi==0.104.1
uvicorn==0.24.0.post1

# 数据库
pymysql==1.1.0
sqlalchemy==2.0.23
pymongo==4.6.1

# 简历解析
python-docx==1.1.0
pdfplumber==0.10.2
pytesseract==0.3.10
opencv-python==4.8.1.78

# 大语言模型/语音
requests==2.31.0
baidu-aip==4.16.10

# 工具类
python-multipart==0.0.6
passlib==1.7.4
python-jose[cryptography]==3.3.0
python-dotenv==1.0.0